importBuild/schem_to_mc_amulet/requirements.txt
amulet-core